pyproject.toml
[tool.poetry]
name = "snaptrack"
version = "0.1.0"
description = ""
authors = ["Krish Shah <shahkrish2016@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.6.1"
python-dotenv = "^1.0.0"
boto3 = "^1.34.11"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
